refactor: minor issues
