+++ src/bin/bitmaskd.rs
@@ -16,8 +16,8 @@ use axum::{
 use bitcoin_30::secp256k1::{ecdh::SharedSecret, PublicKey, SecretKey};
 use bitmask_core::{
     rgb::{
-        accept_transfer, create_invoice, create_psbt, import, issue_contract, list_contracts,
-        list_interfaces, list_schemas, pay_asset,
+        accept_transfer, create_invoice, create_psbt, import as rgb_import, issue_contract,
+        list_contracts, list_interfaces, list_schemas, pay_asset,
     },
     structs::{
         AcceptRequest, ImportRequest, InvoiceRequest, IssueRequest, PsbtRequest, RgbTransferRequest,
@@ -146,14 +146,14 @@ async fn schemas(
     Ok((StatusCode::OK, Json(schemas_res)))
 }
 
-async fn imports(
+async fn import(
     TypedHeader(auth): TypedHeader<Authorization<Bearer>>,
     Json(import_req): Json<ImportRequest>,
 ) -> Result<impl IntoResponse, AppError> {
     info!("POST /accept {import_req:?}");
 
     let nostr_hex_sk = auth.token();
-    let import_res = import(nostr_hex_sk, import_req).await?;
+    let import_res = rgb_import(nostr_hex_sk, import_req).await?;
 
     Ok((StatusCode::OK, Json(import_res)))
 }
@@ -217,7 +217,7 @@ async fn main() -> Result<()> {
         .route("/contracts", get(contracts))
         .route("/interfaces", get(interfaces))
         .route("/schemas", get(schemas))
-        .route("/imports", post(imports))
+        .route("/import", post(import))
         .route("/key/:pk", get(key))
         .route("/carbonado/:pk/:name", post(co_store))
         .route("/carbonado/:pk/:name", get(co_retrieve))
