+++ src/rgb/issue.rs
@@ -1,9 +1,8 @@
+use std::str::FromStr;
+
 use rgbstd::resolvers::ResolveHeight;
 use rgbstd::validation::ResolveTx;
-use std::str::FromStr;
-use strict_types::{svstr, svstruct, StrictVal};
 
-use super::schemas::{default_fungible_iimpl, default_fungible_schema};
 use amplify::confinement::U16;
 use bp::seals::txout::ExplicitSeal;
 use rgbstd::containers::Contract;
@@ -12,6 +11,9 @@ use rgbstd::interface::{BuilderError, ContractBuilder, Iface};
 use rgbstd::persistence::{Inventory, Stash, Stock};
 use rgbstd::Txid;
 use strict_types::encoding::TypeName;
+use strict_types::{svstr, svstruct, StrictVal};
+
+use super::schemas::{default_fungible_iimpl, default_fungible_schema};
 
 #[derive(Clone, Eq, PartialEq, Debug, Display, Error, From)]
 #[display(doc_comments)]
