+++ src/rgb/resolvers.rs
@@ -14,8 +14,7 @@ use rgb::{
     prelude::{DeriveInfo, MiningStatus},
     Utxo,
 };
-use rgbstd::resolvers::ResolveHeight;
-use rgbstd::validation::ResolveTx as ResolveCommiment;
+use rgbstd::{resolvers::ResolveHeight, validation::ResolveTx as ResolveCommiment};
 use wallet::onchain::{ResolveTx, TxResolverError};
 
 pub struct ExplorerResolver {
